+++ packages/@css-blocks/core/test/block-interface-test.ts
@@ -31,7 +31,7 @@ export class BlockInterfaceTests extends BEMProcessor {
     );
 
     let filename = "foo/bar/implements.css";
-    let inputCSS = `@block-reference base from "./base.css";
+    let inputCSS = `@block base from "./base.css";
                     :scope { implements: base; color: red; }
                     .foo { clear: both; }
                     .b[state|small] {color: blue;}`;
@@ -63,8 +63,8 @@ export class BlockInterfaceTests extends BEMProcessor {
     );
 
     let filename = "foo/bar/implements.css";
-    let inputCSS = `@block-reference base from "./base.css";
-                    @block-reference other from "./other.css";
+    let inputCSS = `@block base from "./base.css";
+                    @block other from "./other.css";
                     :scope { implements: base, other; color: red; }
                     .foo { clear: both; }
                     .b[state|small] {color: blue;}
