+++ packages/@css-blocks/ember-cli/README.md
@@ -24,7 +24,7 @@ There are a number of caveats you should know:
  1. Eager engines are a little hacky. We'll want to make some changes to engines proper to make their behavior more consistent with regular addons.
  2. Lazy engines are not possible at this time and will require non-trivial work in in `ember-engines` to support. We have a failing (skipped) test for them in this addon.
  3. The `{{link-to}}` helper (and possibly other built-in form helpers) will require some special casing in `@css-blocks/glimmer` to have a natural feeling integration for basic styling and active states. I have failing (skipped) tests in the project.
- 4. `node_module` resolutions for `@block-reference`s do not work yet.
+ 4. `node_module` resolutions for `@block`s do not work yet.
  5. The optimizer can not be enabled at this time and is hard-coded to disabled.
  6. The broccoli plugins do not have any kind of caching strategy right now! That, and build impact stats / tests, are coming soon.
