+++ packages/@css-blocks/glimmer/test/fixtures/styled-app/src/ui/components/with-multiple-blocks/stylesheet.css
@@ -1,4 +1,4 @@
-@block-reference h from "header.css";
+@block h from "header.css";
 
 :scope {
     color: red;
