+++ packages/@css-blocks/webpack/test/fixtures/blocks/has-reference.block.css
@@ -1,2 +1,2 @@
-@block-reference ref from "is-referenced.block.css";
+@block ref from "is-referenced.block.css";
 @block-debug ref to comment;
